update (v0.15) pre-release 3 utilities, cleanup, layout change
