+++ Source/HMC_NobilityExpanded/HarmonyPermitTablePatch.cs
@@ -70,7 +70,7 @@ namespace NobilityExpanded
     public static class WindowSizePatch
     {
         public static bool Prefix(ref Vector2 __result) {
-            __result = new Vector2(1050f, 880f);
+            __result = new Vector2(1100f, 880f);
             return false;
         }
     }
