--- Source/HMC_NobilityExpanded/NE_Data.cs
@@ -1,16 +1,0 @@
-﻿using UnityEngine;
-
-namespace NobilityExpanded
-{
-    public class NE_Data
-    {
-        public static Vector2 permitOptionSpacing = new Vector2(0.25f, 0.35f);
-        
-        public static string permitCategory = "PermitCategory_";
-        public static string stuffPostfix = "Stuff";
-        public static string spacing = "          ";
-        
-        public string coordsTable = "CoordsTable";
-        public string curTab = "Resources";
-    }
-}
