+++ Source/HMC_NobilityExpanded/NE_Defs/RoyaltyPermitCategoryTableDef.cs
@@ -5,8 +5,6 @@ namespace NobilityExpanded
 {
     public class RoyaltyPermitCategoryTableDef : Def
     {
-        //public List<RoyaltyCoordsTableDef> categories;
         public List<string> categories;
-        
     }
 }
