+++ Source/HMC_NobilityExpanded/PermitsCardCustomUtility.cs
@@ -15,7 +15,7 @@ namespace NobilityExpanded
         public static Faction selectedFaction;
         
         private static Vector2 rightScrollPosition;
-        private static readonly NE_Data Data = new NE_Data();
+        private static string curTab = "Resources";
         private static readonly Texture2D SwitchFactionIcon = ContentFinder<Texture2D>.Get("UI/Icons/SwitchFaction");
         
         private static bool ShowSwitchFactionButton
@@ -100,10 +100,10 @@ namespace NobilityExpanded
             var statusLabelText = "CurrentTitleCustom".Translate() +
                 (currentTitle != null
                     ? currentTitle.GetLabelFor(pawn).CapitalizeFirst()
-                    : (string)"None".Translate()) + NE_Data.spacing + "UnusedPermits".Translate() + ": " +
+                    : (string)"None".Translate()) + Utilities.VarsExposer.spacing + "UnusedPermits".Translate() + ": " +
                 pawn.royalty.GetPermitPoints(selectedFaction).ToString();
             if (!selectedFaction.def.royalFavorLabel.NullOrEmpty())
-                statusLabelText +=  NE_Data.spacing + selectedFaction.def.royalFavorLabel.CapitalizeFirst() + ": " + pawn.royalty.GetFavor(selectedFaction).ToString();
+                statusLabelText +=  Utilities.VarsExposer.spacing + selectedFaction.def.royalFavorLabel.CapitalizeFirst() + ": " + pawn.royalty.GetFavor(selectedFaction).ToString();
             Widgets.Label(statusLabelRect, statusLabelText);
             
             var chooseCategoryRect = new Rect(rect.xMin, rect.y - 4f, 180f, 30f);
@@ -115,11 +115,11 @@ namespace NobilityExpanded
             };
             var middleRect = new Rect(rect) {
                 xMin = 270f,
-                xMax = 790f
+                xMax = 840f
             };
             var rightRect = new Rect(rect) {
-                xMin = 810f,
-                xMax = 1030f
+                xMin = 860f,
+                xMax = 1070f
             };
             
             DoLeftRect(leftRect, pawn);
@@ -136,8 +136,7 @@ namespace NobilityExpanded
         {
             var leftPanel = new Rect(rect);
             Widgets.BeginGroup(leftPanel);
-            if (selectedPermit != null)
-            {
+            if (selectedPermit != null) {
                 Text.Font = GameFont.Medium;
                 var titleRect = new Rect(0.0f, 0.0f, leftPanel.width, 0.0f);
                 Widgets.LabelCacheHeight(ref titleRect, selectedPermit.LabelCap);
@@ -176,21 +175,16 @@ namespace NobilityExpanded
             var defsListForReading = DefDatabase<RoyalTitlePermitDef>.AllDefsListForReading;
             var outRect = rect.ContractedBy(10f);
             var middlePanel = new Rect();
-            // todo foreach optimization?
-            foreach (var permit in defsListForReading.Where(CanDrawPermit)) {
-                middlePanel.width = permit.permitPointCost == 98 ? 
-                    Mathf.Max(middlePanel.width, (float)(DrawPosition(permit).x + 150.0 + 26.0)) : 
-                    Mathf.Max(middlePanel.width, (float)(DrawPosition(permit).x + 200.0 + 26.0));
-                middlePanel.height = Mathf.Max(middlePanel.height, (float)(DrawPosition(permit).y + 50.0 + 26.0));
+            var newCoords = new Vector2(240f, 50f);
+            middlePanel.width = (newCoords + newCoords * new Vector2(0.25f, 0.35f)).x + 200f + 26f;
+            foreach (var permit in defsListForReading.Where(CanDrawPermit).Where(IsFromCurrentCategory)) {
+                middlePanel.height = Mathf.Max(middlePanel.height, DrawPosition(permit).y + 50f + 26f);
             }
 
             Widgets.BeginScrollView(outRect, ref rightScrollPosition, middlePanel);
             Widgets.BeginGroup(middlePanel.ContractedBy(10f));
             DrawLines();
-            foreach (var permit in defsListForReading) {
-                if (!CanDrawPermit(permit) || !IsFromCurrentCategory(permit)) 
-                    continue;
-
+            foreach (var permit in defsListForReading.Where(CanDrawPermit).Where(IsFromCurrentCategory)) {
                 var drawPosition = DrawPosition(permit);
                 var textColor = Widgets.NormalOptionColor;
                 var bgColor = PermitUnlocked(permit, pawn) ? TexUI.FinishedResearchColor : TexUI.AvailResearchColor;
@@ -200,12 +194,12 @@ namespace NobilityExpanded
                     case 99:
                         textColor = new Color(1f, 0.5f, 0.0f, 1.0f);
                         bgColor = new Color(0.06f, 0.06f, 0.06f, 1);
-                        permitRect = new Rect(drawPosition.x, drawPosition.y, 300f, 50f);
+                        permitRect = new Rect(drawPosition.x - 20f, drawPosition.y, 340f, 50f);
                         break;
                     case 98:
                         textColor = new Color(1f, 0.8f, 0.2f, 1.0f);
                         bgColor = new Color(0.06f, 0.06f, 0.06f, 1);
-                        permitRect = new Rect(drawPosition.x - 25f, drawPosition.y, 200f, 50f);
+                        permitRect = new Rect(drawPosition.x - 35f, drawPosition.y, 220f, 50f);
                         break;
                     default:
                         if (!permit.AvailableForPawn(pawn, selectedFaction) && !PermitUnlocked(permit, pawn)) {
@@ -250,13 +244,13 @@ namespace NobilityExpanded
 
                 if (selectedPermit.permitPointCost < 90)
                 {
-                    label += NE_Utility.FormBasicPermitInfo(selectedPermit, selectedFaction, currentTitle, PermitUnlocked(selectedPermit.prerequisite, pawn));
+                    label += Utilities.TextFormatter.FormBasicPermitInfo(selectedPermit, selectedFaction, currentTitle, PermitUnlocked(selectedPermit.prerequisite, pawn));
                     var stuffDefOrdered = DefDatabase<OrderedStuffDef>.GetNamedSilentFail(selectedPermit.defName + "Stuff");
                     bool isStuff = stuffDefOrdered != null;
                     bool isRoyalAidExists = selectedPermit?.royalAid?.itemsToDrop != null;
                     label += isStuff && isRoyalAidExists ? 
-                        NE_Utility.FormAdditionalPermitInfoStuff(selectedPermit) : 
-                        NE_Utility.FormAdditionalPermitInfoExtension(selectedPermit);
+                        Utilities.TextFormatter.FormAdditionalPermitInfoStuff(selectedPermit) : 
+                        Utilities.TextFormatter.FormAdditionalPermitInfoExtension(selectedPermit);
                 }
                 
                 Widgets.LabelCacheHeight(ref infoRect, label);
@@ -289,16 +283,71 @@ namespace NobilityExpanded
                         Widgets.DrawLine(start, end, TexUI.DefaultLineResearchColor, 2f);
                 }
         }
+
+        private static Vector2 DrawPosition(RoyalTitlePermitDef permit){
+            OrderedStuffDef stuffDefOrdered = DefDatabase<OrderedStuffDef>.GetNamedSilentFail(permit.defName + Utilities.VarsExposer.stuffPostfix);
+            bool isExtExists = permit.HasModExtension<PermitExtensionList>();
+            var tableName = Utilities.VarsExposer.coordsTable + curTab + "_";
+            int index;
+            Vector2 newCoords;
+            if (stuffDefOrdered != null) {
+                var categoryTable = DefDatabase<RoyaltyCoordsTableDef>.GetNamedSilentFail(tableName + stuffDefOrdered.column);
+                index = categoryTable.loadOrder.IndexOf(permit);
+                newCoords = new Vector2(categoryTable.coordX * 240f, index * 50f);
+            } else if (isExtExists) {
+                PermitExtensionList extension = permit.GetModExtension<PermitExtensionList>();
+                var column = extension.column ?? 0;
+                var row = extension.row ?? 0;
+                switch (extension.type) {
+                    case "Title":
+                        newCoords = new Vector2(80f, row * 50f + 5f);
+                        break;
+                    case "Category":
+                        newCoords = new Vector2(140f, row * 50f + 5f);
+                        break;
+                    case "Permit":
+                        newCoords = new Vector2(column * 240f, row * 50f);
+                        break;
+                    default:
+                        Log.Error("Error in permit with defName " + permit.defName + " - wrong permit type");
+                        newCoords = new Vector2(-200f, -200f);
+                        break;
+                }
+            } else {
+                RoyaltyCoordsTableDef categoryTable = DefDatabase<RoyaltyCoordsTableDef>.GetNamedSilentFail(tableName + "0");
+                index = categoryTable.loadOrder.IndexOf(permit);
+                switch (permit.permitPointCost)
+                {
+                    case 99:
+                        newCoords = new Vector2(80f, index * 50f + 5f);
+                        break;
+                    case 98:
+                        newCoords = new Vector2(140f, index * 50f + 5f);
+                        break;
+                    default:
+                        newCoords = new Vector2(permit.uiPosition.x * 400f, permit.uiPosition.y * 50f);
+                        break;
+                }
+            }
+            
+            return newCoords + newCoords * new Vector2(0.25f, 0.35f);
+        }
+        
+        private static bool CanDrawPermit(RoyalTitlePermitDef permit) {
+            // todo
+            if (permit.permitPointCost <= 0 || permit.permitPointCost == 97 || permit.permitPointCost == 96)
+                return false;
+            return permit.faction == null || permit.faction == selectedFaction.def;
+        }
         
         private static bool PermitUnlocked(RoyalTitlePermitDef permit, Pawn pawn)
         {
-            if (pawn.royalty.HasPermit(permit, selectedFaction)) return true;
+            if (pawn.royalty.HasPermit(permit, selectedFaction)) 
+                return true;
 
             var allFactionPermits = pawn.royalty.AllFactionPermits;
-            foreach (var factionPermit in allFactionPermits)
-            {
-                if (factionPermit.Permit.prerequisite == permit &&
-                    factionPermit.Faction == selectedFaction)
+            foreach (var factionPermit in allFactionPermits) {
+                if (factionPermit.Permit.prerequisite == permit && factionPermit.Faction == selectedFaction)
                     return true;
                 
                 var secondPrerequisite = factionPermit.Permit.prerequisite?.prerequisite;
@@ -309,68 +358,13 @@ namespace NobilityExpanded
             return false;
         }
 
-        private static Vector2 DrawPosition(RoyalTitlePermitDef permit){
-            OrderedStuffDef stuffDefOrdered = DefDatabase<OrderedStuffDef>.GetNamedSilentFail(permit.defName + NE_Data.stuffPostfix);
-            bool isExtExists = permit.HasModExtension<PermitExtensionList>();
-            int index;
-            Vector2 newCoords;
-            if (stuffDefOrdered != null) {
-                var categoryTable = DefDatabase<RoyaltyCoordsTableDef>.GetNamedSilentFail(Data.coordsTable + Data.curTab + "_" + stuffDefOrdered.column);
-                index = categoryTable.loadOrder.IndexOf(permit);
-                newCoords = new Vector2(categoryTable.coordX * 200f, index * 50f);
-            } else if (isExtExists) {
-                PermitExtensionList extension = permit.GetModExtension<PermitExtensionList>();
-                var column = extension.column ?? 0;
-                var row = extension.row ?? 0;
-                switch (extension.type) {
-                    case "Title":
-                        newCoords = new Vector2(60f, row * 50f + 5f);
-                        break;
-                    case "Category":
-                        newCoords = new Vector2(120f, row * 50f + 5f);
-                        break;
-                    case "Permit":
-                        newCoords = new Vector2(column * 200f, row * 50f);
-                        break;
-                    default:
-                        Log.Error("Error in permit with defName " + permit.defName + " - wrong permit type");
-                        newCoords = new Vector2(-200f, -200f);
-                        break;
-                }
-            } else {
-                RoyaltyCoordsTableDef categoryTable = DefDatabase<RoyaltyCoordsTableDef>.GetNamedSilentFail(Data.coordsTable + Data.curTab + "_0");
-                index = categoryTable.loadOrder.IndexOf(permit);
-                switch (permit.permitPointCost)
-                {
-                    case 99:
-                        newCoords = new Vector2(60f, index * 50f + 5f);
-                        break;
-                    case 98:
-                        newCoords = new Vector2(120f, index * 50f + 5f);
-                        break;
-                    default:
-                        newCoords = new Vector2(permit.uiPosition.x * 400f, permit.uiPosition.y * 50f);
-                        break;
-                }
-            }
-            
-            return newCoords + newCoords * new Vector2(0.25f, 0.35f);;
-        }
-        
-        private static bool CanDrawPermit(RoyalTitlePermitDef permit)
-        {
-            if (permit.permitPointCost <= 0 || permit.permitPointCost == 97 || permit.permitPointCost == 96)
-                return false;
-            return permit.faction == null || permit.faction == selectedFaction.def;
-        }
-
         private static bool IsFromCurrentCategory(RoyalTitlePermitDef permit) {
             var ext = permit.GetModExtension<PermitExtensionList>();
             if (ext != null) {
-                return ext.category == Data.curTab;
+                return ext.category == curTab;
             }
 
-            if (DefDatabase<OrderedStuffDef>.GetNamedSilentFail(permit.defName + NE_Data.stuffPostfix) != null) {
+            if (DefDatabase<OrderedStuffDef>.GetNamedSilentFail(permit.defName + Utilities.VarsExposer.stuffPostfix) != null) {
                 return true;
             }
 
@@ -383,7 +377,7 @@ namespace NobilityExpanded
             var catTable = DefDatabase<RoyaltyPermitCategoryTableDef>.GetNamed("PermitCategoryTable");
             var dict = new Dictionary<string, Action> {};
             foreach (var category in catTable.categories) {
-                dict.Add(NE_Data.permitCategory + category, () => Data.curTab = category);
+                dict.Add(Utilities.VarsExposer.permitCategory + category, () => curTab = category);
             }     
             
             foreach (var table in dict) {
